requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
redis==7.3.0
passlib==1.7.4
python-jose[cryptography]==3.5.0
pydantic-settings==2.13.1
python-multipart==0.0.22
email-validator==2.3.0
pytest==9.0.2
httpx==0.28.1
coverage==7.13.4
locust==2.43.3
